- source clean up

+++ onebusaway-nyc-vehicle-tracking/src/main/java/org/onebusaway/nyc/vehicle_tracking/impl/inference/JourneyPhaseSummaryLibrary.java
@@ -149,7 +149,7 @@ public class JourneyPhaseSummaryLibrary {
     ScheduledBlockLocation blockLocation = blockState.getBlockLocation();
     BlockConfigurationEntry blockConfig = blockInstance.getBlock();
 
-    return (double) blockLocation.getDistanceAlongBlock()
+    return blockLocation.getDistanceAlongBlock()
         / blockConfig.getTotalBlockDistance();
   }
